implemento de eliminar persona

+++ src/main/java/services/InscripcionesPersonas.java
@@ -28,7 +28,7 @@ public class InscripcionesPersonas {
         }
     }
 
-    public void eliminar(int idInscripcion) {
+    public void eliminar(double idInscripcion) {
         listado.removeIf(p -> p.getId() == idInscripcion);
     }
